+++ cli-tests/src/helper.ts
@@ -1,7 +1,12 @@
 import * as fs from 'fs';
 import { spawnSync } from "child_process";
 import * as os from 'os';
+import * as tmp from 'tmp';
+import { paths } from './entities';
+import * as path from 'path';
 
+
+tmp.setGracefulCleanup();
 
 export function executeCommand(command: string, args: string[]) {
   const result = spawnSync(command, args, { encoding: 'utf-8', shell: true, stdio: 'pipe' });
@@ -21,30 +26,11 @@ export const isFileEmpty = (file: string): boolean  => {
     }
 };
 
-export const createDirectory = (file: string): boolean => {
-  try {
-    fs.mkdirSync(file);
-    return true;
-  } catch (error) {
-    return false;
-  }
-};
-
-export const removeDirectory = (file: string): boolean => {
-  if (fs.existsSync(file)) {
-    const stats = fs.statSync(file);
-    if (stats.isDirectory()) {
-      try {
-        fs.rmSync(file, { recursive: true });
-        return true;
-      } catch (error) {
-      }
-    } else {
-      return false;
-    }
-  } else {
-    return false;
-  }
+export const createTmpDirectory = (name = 'tmp-XXXXXX'): tmp.DirResult => {
+  if (!fs.existsSync(paths.pathToOutputDir)) {
+    fs.mkdirSync(paths.pathToOutputDir, { recursive: true });
+  } 
+  return tmp.dirSync({ template: name, tmpdir: paths.pathToOutputDir, unsafeCleanup: true });
 };
 
 export const changeDirectoryPermissions = (directoryPath: string, permission: string): void => {
@@ -79,3 +65,26 @@ export const changeDirectoryPermissions = (directoryPath: string, permission: st
     console.error(`Error changing permissions for directory '${directoryPath}':`, error);
   }
 };
+
+export const pathToSolBinOutputFile = (destination: string): string  => {
+  return path.join(destination, paths.contractSolFilename + paths.binExtension);
+};
+
+export const pathToSolAsmOutputFile = (destination: string): string  => {
+  return path.join(destination, paths.contractSolFilename + paths.asmExtension);
+};
+
+export const pathToLlvmContractsFile = (destination: string): string  => {
+  return path.join(destination, paths.contractLlvmFilename + paths.llvmExtension);
+};
+
+export const isOutputTheSame = (file: string, cliOutput: string): boolean => {
+    const fileOutput = fs.readFileSync(file, 'utf-8');
+    console.log("fileOutput: \n" + fileOutput);
+    console.log("cliOutput: \n" + cliOutput);
+    if (cliOutput.includes(fileOutput) || fileOutput.includes(cliOutput)) {
+        return true
+    } else {
+        return false
+    }
+}