+++ cli-tests/tests/common.test.ts
@@ -1,4 +1,4 @@
-import {executeCommand, isDestinationExist, isFileEmpty} from "../src/helper";
+import {executeCommand, isDestinationExist, isFileEmpty, createTmpDirectory, pathToSolBinOutputFile, pathToSolAsmOutputFile, isOutputTheSame} from "../src/helper";
 import { paths } from '../src/entities';
 
 
@@ -27,12 +27,13 @@ describe("Common tests", () => {
 
     //#1713
     describe(`Default run of ${zksolcCommand} from the help`, () => {
+        const tmpDirZkSolc = createTmpDirectory();
         const args = [
             `"${paths.pathToBasicSolContract}"`,
             `-O3`,
             `--bin`,
             `--output-dir`,
-            `"${paths.pathToOutputDir}"`
+            `"${tmpDirZkSolc.name}"`
         ]; // potential issue on zksolc with full path on Windows cmd
         const result = executeCommand(zksolcCommand, args);
         
@@ -46,31 +47,33 @@ describe("Common tests", () => {
         });
 
         it("Output dir is created", () => {
-            expect(isDestinationExist(paths.pathToOutputDir)).toBe(true);
+            expect(isDestinationExist(tmpDirZkSolc.name)).toBe(true);
         });
 
         xit("Output file is created", () => { // a bug on windows
-            expect(isDestinationExist(paths.pathToSolBinOutputFile)).toBe(true);
+            expect(isDestinationExist(pathToSolBinOutputFile(tmpDirZkSolc.name))).toBe(true);
         });
 
         it("the output file is not empty", () => {
-            expect(isFileEmpty(paths.pathToSolBinOutputFile)).toBe(false);
+            expect(isFileEmpty(pathToSolBinOutputFile(tmpDirZkSolc.name))).toBe(false);
         });
 
         it("No 'Error'/'Warning'/'Fail' in the output", () => {
             expect(result.output).not.toMatch(/([Ee]rror|[Ww]arning|[Ff]ail)/i);
+            tmpDirZkSolc.removeCallback();
         });
     });
 
     //#1818
     describe(`Run ${zksolcCommand} with multiple output options from the help`, () => {
+        const tmpDirZkSolc = createTmpDirectory();
         const args = [
             `"${paths.pathToBasicSolContract}"`,
             `-O3`,
             `--bin`,
             `--asm`,
             `--output-dir`,
-            `"${paths.pathToOutputDir}"`
+            `"${tmpDirZkSolc.name}"`
         ]; // potential issue on zksolc with full path on Windows cmd
         const result = executeCommand(zksolcCommand, args);
 
@@ -81,18 +84,55 @@ describe("Common tests", () => {
             expect(result.exitCode).toBe(0);
         });
         it("Output dir is created", () => {
-            expect(isDestinationExist(paths.pathToOutputDir)).toBe(true);
+            expect(isDestinationExist(tmpDirZkSolc.name)).toBe(true);
         });
         xit("Output files are created", () => { // a bug on windows
-            expect(isDestinationExist(paths.pathToSolBinOutputFile)).toBe(true);
-            expect(isDestinationExist(paths.pathToSolAsmOutputFile)).toBe(true);
+            expect(isDestinationExist(pathToSolBinOutputFile(tmpDirZkSolc.name))).toBe(true);
+            expect(isDestinationExist(pathToSolAsmOutputFile(tmpDirZkSolc.name))).toBe(true);
         });
         it("the output files are not empty", () => {
-            expect(isFileEmpty(paths.pathToSolBinOutputFile)).toBe(false);
-            expect(isFileEmpty(paths.pathToSolAsmOutputFile)).toBe(false);
+            expect(isFileEmpty(pathToSolBinOutputFile(tmpDirZkSolc.name))).toBe(false);
+            expect(isFileEmpty(pathToSolAsmOutputFile(tmpDirZkSolc.name))).toBe(false);
         });
         it("No 'Error'/'Warning'/'Fail' in the output", () => {
             expect(result.output).not.toMatch(/([Ee]rror|[Ww]arning|[Ff]ail)/i);
+            tmpDirZkSolc.removeCallback();
+        });
+    });
+
+    //#issue CPR-1498 Different --bin output for a file and cli
+    describe(`--bin output is the same for in a file and in a cli`, () => {
+        const tmpDirZkSolc = createTmpDirectory();
+        const args = [
+            `"${paths.pathToBasicSolContract}"`,
+            `-O3`,
+            `--bin`,
+            `--output-dir`,
+            `"${tmpDirZkSolc.name}"`
+        ]; // potential issue on zksolc with full path on Windows cmd
+        const result = executeCommand(zksolcCommand, args);
+        
+        it("Compiler run successful", () => {
+            expect(result.output).toMatch(/(Compiler run successful.)/i);
+        });
+
+        it("Exit code = 0", () => {
+            expect(result.exitCode).toBe(0);
+        });
+
+        it("Output file is created", () => { // a bug on windows
+            expect(isDestinationExist(pathToSolBinOutputFile(tmpDirZkSolc.name))).toBe(true);
+        });
+
+        it("Output file == cli output", () => { // a bug on windows
+            const args = [
+                `"${paths.pathToBasicSolContract}"`,
+                `-O3`,
+                `--bin`
+            ]; // potential issue on zksolc with full path on Windows cmd
+            const resultCli = executeCommand(zksolcCommand, args);
+            expect(isOutputTheSame(pathToSolBinOutputFile(tmpDirZkSolc.name), resultCli.output)).toBe(true);
+            tmpDirZkSolc.removeCallback();
         });
     });
 });