+++ src/solc/pipeline.rs
@@ -9,8 +9,6 @@ use crate::solc::Compiler as SolcCompiler;
 /// The Solidity compiler pipeline type.
 ///
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[allow(non_camel_case_types)]
-#[allow(clippy::upper_case_acronyms)]
 pub enum Pipeline {
     /// The Yul IR.
     Yul,