+++ src/solc/standard_json/input/settings/selection/file/flag.rs
@@ -11,8 +11,6 @@ use crate::solc::pipeline::Pipeline as SolcPipeline;
 /// The `solc --standard-json` expected output selection flag.
 ///
 #[derive(Debug, Serialize, Deserialize, Clone, Copy, PartialEq, Eq, Hash)]
-#[allow(non_camel_case_types)]
-#[allow(clippy::upper_case_acronyms)]
 pub enum Flag {
     /// The ABI JSON.
     #[serde(rename = "abi")]